requirements.txt
telethon==1.36.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
selenium==4.1.0
webdriver-manager==4.0.1
yadisk==2.1.0
PyYAML==6.0.1